# Conflicts:
#	app/src/main/java/com/roacult/kero/oxxy/projet2eme/ui/start_chalenge/StartChelngeViewModel.kt
#	domain/src/main/java/com/roacult/kero/oxxy/domain/interactors/GetChalengeDetaills.kt
#	domain/src/main/java/com/roacult/kero/oxxy/domain/interactors/SetUserTry.kt

+++ data/src/main/java/com/roacult/kero/oxxy/projet2eme/network/entities/GetAllChallengeReponse.kt
@@ -3,4 +3,4 @@ package com.roacult.kero.oxxy.projet2eme.network.entities
 import com.roacult.kero.oxxy.domain.modules.ChalengeGlobale
 
 data class GetAllChallengeReponse(val reponse:Int, val challenges: List<ChalengeGlobale>?)
-data class Request( val year :Int , val id:Long)
+data class Request( val year :Int , val id:Int)
